change folders structure root
