--- src/Assets/NavBar.elements.js
@@ -1,191 +1,0 @@
-// import styled from "styled-components";
-import styled from "styled-components";
-import { Link } from "react-router-dom";
-
-export const Container = styled.div`
-  width: 100%;
-  height: 70px;
-  //   background-color: #800;
-  position: relative;
-`;
-
-export const Wrapper = styled.div`
-  width: 100%;
-  max-width: 1300px;
-  height: 100%;
-  display: flex;
-  justify-content: space-between;
-  //   margin: auto;
-`;
-
-export const LogoContainer = styled.div`
-  margin-left: 0.5rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  //   background: aqua;
-  height: 100%;
-
-  path {
-    fill: #283159;
-    width: 10;
-  }
-`;
-
-export const Menu = styled.ul`
-  height: 100%;
-  display: flex;
-  justify-content: space-between;
-  list-style: none;
-  margin-left: 30px;
-  @media (max-width: 768px) {
-    display: none;
-  }
-`;
-
-// @media screen and (max-width: 960px) {
-export const DropdownContainer = styled.div`
-  position: absolute;
-  top: 0px;
-  left: 0;
-  justify-content: center;
-  flex-direction: column;
-  align-items: center;
-  transition: 0.5s all ease;
-  background: green;
-  z-index: 1;
-  padding: 1rem;
-  background: #ffefc0;
-  width: 65%;
-
-  @media (min-width: 768px) {
-    display: none;
-  }
-`;
-
-export const DropdownHeader = styled.div`
-  display: flex;
-  justify-content: flex-end;
-`;
-
-export const DropdownMenu = styled.div`
-  height: 250px;
-  margin-top: 80px;
-  display: flex;
-  flex-direction: column;
-`;
-
-export const DropdownItem = styled.div`
-  height: 50px;
-  display: flex;
-  align-items: center;
-  /* justify-content: center; */
-  transition: background 0.5s;
-  padding: 0.5rem;
-`;
-
-export const MenuItem = styled.li`
-  height: 100%;
-`;
-
-// @media screen and (max-width: 960px) {
-export const BurgerIconContainer = styled.div`
-  display: none;
-
-  @media (max-width: 768px) {
-    align-items: center;
-    // background: aqua;
-    cursor: pointer;
-    margin-left: 0.5rem;
-    display: flex;
-    justify-content: center;
-    height: 100%;
-
-    path {
-      fill: #283159;
-      margin-right: 0.5rem;
-    }
-  }
-`;
-
-export const UserContainer = styled.div`
-  background-color: #ff9aa3;
-  display: flex;
-  border-radius: 100%;
-  justify-content: center;
-  align-items: center;
-  width: 40px;
-  height: 40px;
-
-  path {
-    fill: #fff;
-  }
-`;
-
-export const Block = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-export const NumberCart = styled.span`
-  position: absolute;
-  left: -6px;
-  bottom: -6px;
-  // left: 0;
-  // bottom: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 100%;
-  width: 20px;
-  height: 20px;
-  background: #ff9aa3;
-
-  // color: #283159;
-  color: #fff;
-  font-weight: 700;
-`;
-
-export const StyledLink = styled(Link)`
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  height: 100%;
-  padding: 0.5rem 2.5rem;
-  color: #283159;
-  font-family: sans-serif;
-  font-size: 1rem;
-  font-weight: 300;
-  cursor: pointer;
-  transition: 0.5s all ease;
-  text-decoration: none;
-  font-weight: 700;
-
-  &:hover {
-    color: #ccc;
-    background-color: 283159;
-    transition: 0.5 all ease;
-  }
-`;
-
-export const StyledLinkDropdown = styled(Link)`
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  height: 100%;
-  padding: 0.5rem 2.5rem;
-  color: #283159;
-  font-family: sans-serif;
-  font-size: 1rem;
-  font-weight: 300;
-  cursor: pointer;
-  transition: 0.5s all ease;
-  text-decoration: none;
-  font-weight: 700;
-
-  &:hover {
-    color: #ccc;
-    background-color: 283159;
-    transition: 0.5 all ease;
-  }
-`;
