+++ src/Components/App.js
@@ -7,33 +7,9 @@ import { AuthProvider } from "./Context/AuthContext";
 import { RoleComponent } from "./ProtectedRoutes";
 import ChefView from "./ChefView";
 import WaiterView from "./WaiterView";
-// -----------------------
-import OrdersResumeWaiter from "./Pages/Waiter/OrdersResumeWaiter";
-import TakeOrderWaiter from "./Pages/Waiter/TakeOrderWaiter";
+import OrdersResumeWaiter from "../Pages/Waiter/OrdersResumeWaiter";
+import TakeOrderWaiter from "../Pages/Waiter/TakeOrderWaiter";
 
-// function App() {
-//   return (
-//     <AuthProvider>
-//       <Routes>
-//         <Route
-//           path="/home"
-//           element={
-//             <RoleComponent role="admin">
-//               <Home />
-//             </RoleComponent>
-//           }
-//         ></Route>
-
-//         {/* Waiter Views */}
-//         <Route path="/" element={<Login />}></Route>
-//         <Route path="/chef/*" element={<ChefView />}></Route>
-//         {/* <Route path="/chef" element={<ChefView />}></Route> */}
-//       </Routes>
-//     </AuthProvider>
-//   );
-// }
-
-// ! EL QUE FUNCIONA
 function App() {
   return (
     <AuthProvider>
@@ -100,9 +76,8 @@ function App() {
             </RoleComponent>
           }
         ></Route>
-        {/* <Route path="/Button" element={<ButtonFilter />}></Route> */}
-        {/* Chef Views */}
 
+        {/* Chef Views */}
         <Route
           path="/chef"
           element={
