+++ src/Components/Search.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { getProducts } from "./Context/FirestoreContext";
+import { getProducts } from "./Context/FirestoreServices";
 import { ReactComponent as Look } from "../Assets/icons/magnifying-glass.svg";
 const Search = ({ onChange }) => {
   const handleChange = (e) => {
