+++ src/Pages/Waiter/OrdersResumeWaiter.js
@@ -1,43 +1,32 @@
 import React, { useState, useEffect } from "react";
-import "../../../Assets/Navbar.css";
-import "../../../Assets/Sidebar.css";
-import ProductCart from "../../../Components/ProductCard";
-import ButtonFilter from "../../../Components/ButtonFilter";
-import iconComponents from "../../../Assets/CustomLogo";
+import "../../Assets/Navbar.css";
+import "../../Assets/Sidebar.css";
+import ButtonFilter from "../../Components/ButtonFilter";
+import iconComponents from "../../Assets/CustomLogo";
 import {
   getProducts,
   getProductsCategories,
   filterProductByCategorie,
-} from "../../../Components/Context/FirestoreContext";
-import Search from "../../Search";
+} from "../../Components/Context/FirestoreServices";
+import OrderCardFormat from "../../Components/OrderCardFormat";
 
-const TakeOrderWaiter = () => {
-  const [products, setProducts] = useState([]);
+const OrdersResumeWaiter = () => {
   const [productCategories, setProductCategories] = useState([]);
-
   const handleCategorie = async (catUid, catName) =>
     await filterProductByCategorie(catUid, catName);
   useEffect(() => {
-    getProducts().then((products) => setProducts(products));
+    // getProducts().then((products) => setProducts(products));
     getProductsCategories().then((category) => setProductCategories(category));
   }, []);
 
   const handleClick = ({ cat_uid, cat_name }) => {
     handleCategorie(cat_uid, cat_name).then((items) => {
-      setProducts(items);
+      // setProducts(items);
     });
   };
 
-  const handleSearch = async (query) => {
-    const products = await getProducts();
-    const product = products.filter((elem) => {
-      return elem.product_name.toLowerCase().includes(query.toLowerCase());
-    });
-    setProducts(product);
-  };
   return (
     <>
-      <Search onChange={handleSearch}></Search>
       <div className="categories-container">
         {productCategories.map((cat, i) => {
           return (
@@ -56,13 +45,9 @@ const TakeOrderWaiter = () => {
           );
         })}
       </div>
-      <div className="products-container">
-        {products.map((p) => {
-          return <ProductCart product={p} />;
-        })}
-      </div>
+      <OrderCardFormat />
     </>
   );
 };
 
-export default TakeOrderWaiter;
+export default OrdersResumeWaiter;
